extract helper functions
