+++ include/ck/tensor_operation/gpu/device/device_batched_gemm_softmax_gemm_permute_xdl_cshuffle.hpp
@@ -263,74 +263,17 @@ struct DeviceBatchedGemmSoftmaxGemmPermute_Xdl_CShuffle
     using B1GridDesc_G_N_K     = decltype(Transform::MakeB1GridDescriptor_G_N_K({}, {}));
     using CGridDesc_G_M_N      = decltype(Transform::MakeCGridDescriptor_G_M_N({}, {}));
 
-    struct MaskDisabled
-    {
-        __host__ __device__ constexpr bool operator()(index_t /*m*/, index_t /*n*/) const
-        {
-            return false;
-        };
-
-        __host__ __device__ constexpr bool IsTileSkippable(index_t /*m*/,
-                                                           index_t /*n*/,
-                                                           index_t /*m_tile*/,
-                                                           index_t /*n_tile*/) const
-        {
-            return false;
-        }
-    };
-
-    struct MaskOutUpperTriangle
-    {
-        __host__ __device__ constexpr bool operator()(index_t m, index_t n) const { return n > m; }
-
-        __host__ __device__ constexpr bool
-        IsTileSkippable(index_t m, index_t n, index_t m_tile, index_t /*n_tile*/) const
-        {
-            return operator()(m + m_tile - 1, n);
-        }
-    };
-
     constexpr static auto make_MaskOutPredicate()
     {
         if constexpr(MaskingSpec == MaskingSpecialization::MaskDisabled)
         {
-            return MaskDisabled{};
+            return MaskDisabledPredicate{};
         }
         else if constexpr(MaskingSpec == MaskingSpecialization::MaskOutUpperTriangle)
         {
-            return MaskOutUpperTriangle{};
+            return MaskOutUpperTrianglePredicate{};
         }
     }
-
-    // to track the points which need to be set to -inf on C0
-    // Note: no need to reset M padding value, because they will not be stored out.
-    template <typename MaskOutPredicate>
-    struct C0MatrixMask_impl
-    {
-        C0MatrixMask_impl(index_t NRaw) : NRaw_(NRaw), predicate_(MaskOutPredicate{}) {}
-
-        __host__ __device__ constexpr bool IsNOutOfBound(/*index_t m, */ index_t n) const
-        {
-            return n >= NRaw_;
-        }
-
-        __host__ __device__ constexpr bool IsMaskedElement(index_t m, index_t n) const
-        {
-            return predicate_(m, n) || IsNOutOfBound(n);
-        }
-
-        __host__ __device__ constexpr bool
-        IsTileSkippable(index_t m, index_t n, index_t m_tile, index_t n_tile) const
-        {
-            return predicate_.IsTileSkippable(m, n, m_tile, n_tile);
-        }
-
-        private:
-        // index_t MRaw_;
-        index_t NRaw_;
-        MaskOutPredicate predicate_;
-    };
-
     using C0MatrixMask = C0MatrixMask_impl<decltype(make_MaskOutPredicate())>;
 
     struct ComputeBasePtrOfStridedBatch
